requirements.txt
google-ads>=24.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.2.0
facebook-business>=20.0.0
python-dotenv>=1.0.0
requests>=2.31.0
gspread>=6.0.0
pymupdf>=1.27.0
